Small optimizations (#23)
* unnecessary boxing allocation

* use nameof() over string value

* redundant explicit array type specification

* wrong namespace

* not need to assigned initialization value

* Redundant parenthesis

* casting not needed

* unnecessary using statements

* unnecessary base constructor call

* obj initializer

* redundent else statements

* .ConfigureAwait(false)

* consistency

+++ src/Microsoft.Azure.ServiceBus/Primitives/Ticks.cs
@@ -18,7 +18,7 @@ namespace Microsoft.Azure.ServiceBus
 
         public static long FromMilliseconds(int milliseconds)
         {
-            return checked((long)milliseconds * TimeSpan.TicksPerMillisecond);
+            return checked(milliseconds * TimeSpan.TicksPerMillisecond);
         }
 
         public static int ToMilliseconds(long ticks)
